Displaying retrieved data from the TMDB API on the homepage.
Adding a series in AddNew results in displaying the retrieved data from the TMDB API on the homepage: title and poster. The AddNew page is divided into two sections: searching for data from the API and adding the series to the list. I've also created the design for the AddNew page in Figma.