Fixed bugs and added new bugs as well.
fixed bug: there was a bug with fetch - in the case of the title of the series not being found, the fetch did not treat it as an error, therefore the component informing about it did not render. I added an additional condition to fix this.
Scaffolding for the UserReview component. The initial conditional rendering (pressing the "Dodaj opinię" button causes the user review to render as a <div> instead of a <textarea>.
To do: Pass the user review to parent components without causing rendering errors simultaneously.