Disabled/enabled submitting button
Conditional rendering of the Dodaj button in the AddNew component based on whether the user has added their review and rated the series. Extracting the ShowDetails component (with the series metrics) from the ThereIsResult component. Fixing minor bugs (review state not resetting if the user wrote a review, did not add a series, and then searched for a new title).