Editing the series in the modal works correctly.
The new values of review and userStarRate are correctly passed to the [title] array in Zustand, and then rendered correctly in the TitlesDisplay component.
The StarRatingUser component conditionally renders UI based on the provided props (initialValue and edit).
Possible improvement: When editing the rating, clicking the 'Edytuj' button should default to displaying the previously assigned rating.